+++ experimental/traffic-portal-angular8/src/app/services/api/DeliveryServiceService.ts
@@ -11,20 +11,15 @@
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */
-import { Injectable } from '@angular/core';
-import { HttpClient, HttpHeaders, HttpResponse, HttpParams } from '@angular/common/http';
-import { BehaviorSubject, Observable, throwError } from 'rxjs';
-import { merge } from 'rxjs/index';
-import { map, mergeAll, first, catchError, reduce } from 'rxjs/operators';
 
-import { environment } from '../../environments/environment';
-import { CDN } from '../models/cdn';
-import { DataPoint, DataSet, DataSetWithSummary, TPSData } from '../models/data';
-import { DeliveryService } from '../models/deliveryservice';
-import { InvalidationJob } from '../models/invalidation';
-import { Type } from '../models/type';
-import { Role, User, Capability } from '../models/user';
-import { Server } from '../models/server';
+import { Injectable } from '@angular/core';
+import { Observable } from 'rxjs';
+import { merge } from 'rxjs/index';
+import { map, mergeAll, reduce } from 'rxjs/operators';
+
+import { APIService } from './apiservice';
+
+import { DataPoint, DataSet, DataSetWithSummary, DeliveryService, TPSData, Type } from '../../models';
 
 function constructDataSetFromResponse (r: any): DataSetWithSummary {
 	if (!r.series || !r.series.name) {
@@ -75,88 +70,11 @@ function constructDataSetFromResponse (r: any): DataSetWithSummary {
 	} as DataSetWithSummary;
 }
 
-/**
- * The APIService provides access to the Traffic Ops API. Its methods should be kept API-version
- * agnostic (from the caller's perspective), and always return `Observable`s.
-*/
-@Injectable({ providedIn: 'root' })
-export class APIService {
 
-	/**
-	 * The current API version to use
-	 */
-	public API_VERSION: string = '2.0';
+@Injectable({providedIn: 'root'})
+export class DeliveryServiceService extends APIService {
 
 	private deliveryServiceTypes: Array<Type>;
-
-	// private cookies: string;
-
-	constructor (private readonly http: HttpClient) {
-		if (environment.APIVersion) {
-			this.API_VERSION = environment.APIVersion;
-		}
-	}
-
-	private delete (path: string, data?: any): Observable<HttpResponse<any>> {
-		return this.do('delete', path, data);
-	}
-	private get (path: string, data?: any): Observable<HttpResponse<any>> {
-		return this.do('get', path, data);
-	}
-	private head (path: string, data?: any): Observable<HttpResponse<any>> {
-		return this.do('head', path, data);
-	}
-	private options (path: string, data?: any): Observable<HttpResponse<any>> {
-		return this.do('options', path, data);
-	}
-	private patch (path: string, data?: any): Observable<HttpResponse<any>> {
-		return this.do('patch', path, data);
-	}
-	private post (path: string, data?: any): Observable<HttpResponse<any>> {
-		return this.do('post', path, data);
-	}
-	private push (path: string, data?: any): Observable<HttpResponse<any>> {
-		return this.do('push', path, data);
-	}
-
-	private do (method: string, path: string, data?: Object): Observable<HttpResponse<any>> {
-
-		/* tslint:disable */
-		const options = {headers: new HttpHeaders({'Content-Type': 'application/json'}),
-		                 observe: 'response' as 'response',
-		                 responseType: 'json' as 'json',
-		                 body: data};
-		/* tslint:enable */
-		return this.http.request(method, path, options).pipe(map((response) => {
-			// TODO pass alerts to the alert service
-			// (TODO create the alert service)
-			return response as HttpResponse<any>;
-		}));
-	}
-
-	/**
-	 * Performs authentication with the Traffic Ops server.
-	 * @param u The username to be used for authentication
-	 * @param p The password of user `u`
-	 * @returns An observable that will emit the entire HTTP response
-	*/
-	public login (u: string, p: string): Observable<HttpResponse<any>> {
-		const path = '/api/' + this.API_VERSION + '/user/login';
-		return this.post(path, {u, p});
-	}
-
-	/**
-	 * Fetches the current user from Traffic Ops
-	 * @returns An observable that will emit a `User` object representing the current user.
-	*/
-	public getCurrentUser (): Observable<User> {
-		const path = '/api/' + this.API_VERSION + '/user/current';
-		return this.get(path).pipe(map(
-			r => {
-				return r.body.response as User;
-			}
-		));
-	}
 
 	/**
 	 * Gets a list of all visible Delivery Services
@@ -390,162 +308,6 @@ export class APIService {
 	}
 
 	/**
-	 * Gets an array of all users in Traffic Ops
-	 * @returns An Observable that will emit an Array of User objects.
-	*/
-	public getUsers (nameOrID?: string | number): Observable<Array<User> | User> {
-		const path = '/api/' + this.API_VERSION + '/users';
-		if (nameOrID) {
-			switch (typeof nameOrID) {
-				case 'string':
-					return this.get(path + '?username=' + encodeURIComponent(nameOrID)).pipe(map(
-						r => {
-							for (const u of (r.body.response as Array<User>)) {
-								if (u.username === nameOrID) {
-									return u;
-								}
-							}
-							return null;
-						}
-					));
-				case 'number':
-					return this.get(path + '?id=' + nameOrID.toString()).pipe(map(
-						r => {
-							for (const u of (r.body.response as Array<User>)) {
-								if (u.id === nameOrID) {
-									return u;
-								}
-							}
-							return null;
-						}
-					));
-				default:
-					throw new TypeError("expected a username or ID, got '" + typeof(nameOrID) + "'");
-					return null;
-			}
-		}
-		return this.get(path).pipe(map(
-			r => {
-				return r.body.response as Array<User>;
-			}
-		));
-	}
-
-	public getRoles (id: number): Observable<Role>;
-	public getRoles (name: string): Observable<Role>;
-	public getRoles (): Observable<Array<Role>>;
-	/**
-	 * Fetches one or all Roles from Traffic Ops
-	 * @param name Optionally, the name of a single Role which will be fetched
-	 * @param id Optionally, the integral, unique identifier of a single Role which will be fetched
-	 * @throws {TypeError} When called with an improper argument.
-	 * @returns an Observable that will emit either an Array of Roles, or a single Role, depending on whether
-	 *	`name`/`id` was passed
-	 * (In the event that `name`/`id` is given but does not match any Role, `null` will be emitted)
-	*/
-	public getRoles (nameOrID?: string | number) {
-		const path = '/api/' + this.API_VERSION + '/roles';
-		if (nameOrID) {
-			switch (typeof nameOrID) {
-				case 'string':
-					return this.get(path + '?name=' + nameOrID).pipe(map(
-						r => {
-							for (const role of (r.body.response as Array<Role>)) {
-								if (role.name === nameOrID) {
-									return role;
-								}
-							}
-							return null;
-						}
-					));
-					break;
-				case 'number':
-					return this.get(path + '?id=' + nameOrID.toString()).pipe(map(
-						r => {
-							for (const role of (r.body.response as Array<Role>)) {
-								if (role.id === nameOrID) {
-									return role;
-								}
-							}
-						}
-					));
-					break;
-				default:
-					throw new TypeError("expected a name or ID, got '" + typeof(nameOrID) + "'");
-					break;
-			}
-		}
-		return this.get(path).pipe(map(
-			r => {
-				return r.body.response as Array<Role>;
-			}
-		));
-	}
-
-	public getCapabilities (name: string): Observable<Capability>;
-	public getCapabilities (): Observable<Array<Capability>>;
-	/**
-	 * Fetches one or all Capabilities from Traffic Ops
-	 * @param name Optionally, the name of a single Capability which will be fetched
-	 * @throws {TypeError} When called with an improper argument.
-	 * @returns an Observable that will emit either an Array of Capabilities, or a single Capability,
-	 * depending on whether `name`/`id` was passed
-	 * (In the event that `name`/`id` is given but does not match any Capability, `null` will be emitted)
-	*/
-	public getCapabilities (name?: string) {
-		const path = '/api/' + this.API_VERSION + '/capabilities';
-		if (name) {
-			return this.get(path + '?name=' + encodeURIComponent(name)).pipe(map(
-				r => {
-					for (const cap of (r.body.response as Array<Capability>)) {
-						if (cap.name === name) {
-							return cap;
-						}
-					}
-					return null;
-				}
-			));
-		}
-		return this.get(path).pipe(map(
-			r => {
-				return r.body.response as Array<Capability>;
-			}
-		));
-	}
-
-	/**
-	 * Gets one or all Types from Traffic Ops
-	 * @param name Optionally, the name of a single Type which will be returned
-	 * @returns An Observable that will emit either a Map of Type names to full Type objects, or a single Type, depending on whether
-	 * 	`name` was passed
-	 * (In the event that `name` is given but does not match any Type, `null` will be emitted)
-	*/
-	public getTypes (name?: string): Observable<Map<string, Type> | Type> {
-		const path = '/api/' + this.API_VERSION + '/types';
-		if (name) {
-			return this.get(path + '?name=' + encodeURIComponent(name)).pipe(map(
-				r => {
-					for (const t of (r.body.response as Array<Type>)) {
-						if (t.name === name) {
-							return t;
-						}
-					}
-					return null;
-				}
-			));
-		}
-		return this.get(path).pipe(map(
-			r => {
-				const ret = new Map<string, Type>();
-				for (const t of (r.body.response as Array<Type>)) {
-					ret.set(t.name, t);
-				}
-				return ret;
-			}
-		));
-	}
-
-	/**
 	 * This method is handled seperately from :js:method:`APIService.getTypes` because this information
 	 * (should) never change, and therefore can be cached. This method makes an HTTP request iff the values are not already
 	 * cached.
@@ -567,81 +329,6 @@ export class APIService {
 			r => {
 				this.deliveryServiceTypes = r.body.response as Array<Type>;
 				return r.body.response as Array<Type>;
-			}
-		));
-	}
-
-	/**
-	 * Gets one or all CDNs from Traffic Ops
-	 * @param id The integral, unique identifier of a single CDN to be returned
-	 * @returns An Observable that will emit either a Map of CDN names to full CDN objects, or a single CDN, depending on whether `id` was
-	 * 	passed.
-	 * (In the event that `id` is passed but does not match any CDN, `null` will be emitted)
-	*/
-	public getCDNs (id?: number): Observable<Map<string, CDN> | CDN> {
-		const path = '/api/' + this.API_VERSION + '/cdns';
-		if (id) {
-			return this.get(path + '?id=' + String(id)).pipe(map(
-				r => {
-					for (const c of (r.body.response as Array<CDN>)) {
-						if (c.id === id) {
-							return c;
-						}
-					}
-				}
-			));
-		}
-		return this.get(path).pipe(map(
-			r => {
-				const ret = new Map<string, CDN>();
-				for (const c of (r.body.response as Array<CDN>)) {
-					ret.set(c.name, c);
-				}
-				return ret;
-			}
-		));
-	}
-
-	public getInvalidationJobs (opts?: {id: number} |
-	                                   {userId: number} |
-	                                   {user: User} |
-	                                   {dsId: number} |
-	                                   {deliveryService: DeliveryService}): Observable<Array<InvalidationJob>> {
-		let path = '/api/' + this.API_VERSION + '/jobs';
-		if (opts) {
-			path += '?';
-			if (opts.hasOwnProperty('id')) {
-				path += 'id=' + String((opts as {id: number}).id);
-			} else if (opts.hasOwnProperty('dsId')) {
-				path += 'dsId=' + String((opts as {dsId: number}).dsId);
-			} else if (opts.hasOwnProperty('userId')) {
-				path += 'userId=' + String((opts as {userId: number}).userId);
-			} else if (opts.hasOwnProperty('deliveryService')) {
-				path += 'dsId=' + String((opts as {deliveryService: DeliveryService}).deliveryService.id);
-			} else {
-				path += 'userId=' + String((opts as {user: User}).user.id);
-			}
-		}
-		return this.get(path).pipe(map(
-			r => {
-				return r.body.response as Array<InvalidationJob>;
-			}
-		));
-	}
-
-	public createInvalidationJob (job: InvalidationJob): Observable<boolean> {
-		const path = '/api/' + this.API_VERSION + '/user/current/jobs';
-		return this.post(path, job).pipe(map(
-			r => true,
-			e => false
-		));
-	}
-
-	public getServers(): Observable<Array<Server>> {
-		let path = `/api/${this.API_VERSION}/servers`;
-		return this.get(path).pipe(map(
-			r => {
-				return r.body.response as Array<Server>;
 			}
 		));
 	}
